+++ packages/lib/test/src/app/App.test.js
@@ -5,20 +5,20 @@ import App from '../../../src/app/App';
 import Contracts from '../../../src/utils/Contracts'
 import Proxy from '../../../src/utils/Proxy';
 
-const AppDirectory = Contracts.getFromLocal('AppDirectory');
 const Impl = Contracts.getFromLocal('Impl');
 const ImplV1 = Contracts.getFromLocal('DummyImplementation');
 const ImplV2 = Contracts.getFromLocal('DummyImplementationV2');
+const AppDirectory = Contracts.getFromLocal('AppDirectory');
+const ImplementationDirectory = Contracts.getFromLocal('ImplementationDirectory');
 
 contract('App', function ([_, owner, otherAdmin]) {
   const txParams = { from: owner }
   const initialVersion = '1.0';
   const contractName = 'Impl';
-  const stdlibAddress = '0x0000000000000000000000000000000000000010';
 
   const shouldInitialize = function () {
     it('deploys all contracts', async function() {
-      this.app.address().should.not.be.null;
+      this.app.address.should.not.be.null;
       this.app.factory.address.should.not.be.null;
       this.app.package.address.should.not.be.null;
     });
@@ -33,19 +33,18 @@ contract('App', function ([_, owner, otherAdmin]) {
 
     it('initializes all app properties', async function () {
       this.app.version.should.eq(initialVersion);
-      this.app.directories.should.have.key(initialVersion);
+      this.app.directory.should.not.be.null
     });
 
     it('returns the current directory', async function () {
-      this.app.currentDirectory().address.should.be.not.null;
+      this.app.directory.address.should.be.not.null;
     });
   };
 
   const shouldConnectToStdlib = function () {
     it('should connect current directory to stdlib', async function () {
-      const address = await this.app.package.getVersion(this.app.version);
-      const currentDirectory = AppDirectory.at(address)
-      const currentStdlib = await currentDirectory.stdlib()
+      const appDirectory = await this.app.package.getDirectory(this.app.version)
+      const currentStdlib = await appDirectory.stdlib()
 
       const stdlib = await this.app.currentStdlib();
       stdlib.should.be.eq(currentStdlib)
@@ -55,6 +54,10 @@ contract('App', function ([_, owner, otherAdmin]) {
       (await this.app.hasStdlib()).should.be.true
     })
   };
+
+  beforeEach('deploying stdlib', async function () {
+    this.stdlib = await ImplementationDirectory.new({ from: owner })
+  });
 
   describe('without stdlib', function () {
     beforeEach('deploying', async function () {
@@ -69,9 +72,9 @@ contract('App', function ([_, owner, otherAdmin]) {
       })
     });
 
-    describe('connect', function () {
+    describe('fetch', function () {
       beforeEach('connecting to existing instance', async function () {
-        this.app = await App.fetch(this.app.address(), txParams);
+        this.app = await App.fetch(this.app.address, txParams);
       });
 
       shouldInitialize();
@@ -91,7 +94,7 @@ contract('App', function ([_, owner, otherAdmin]) {
 
       it('updates own properties', async function () {
         this.app.version.should.eq(newVersion);
-        this.app.directories.should.include.key(newVersion);
+        this.app.directory.should.not.be.null
       });
 
       it('registers new version on package', async function () {
@@ -99,8 +102,9 @@ contract('App', function ([_, owner, otherAdmin]) {
       });
 
       it('returns the current directory', async function () {
-        const currentDirectory = await this.app.package.getVersion(newVersion);
-        this.app.currentDirectory().address.should.eq(currentDirectory);
+        const appDirectory = await this.app.package.getDirectory(this.app.version)
+
+        this.app.directory.address.should.eq(appDirectory.address)
       });
     });
 
@@ -116,7 +120,7 @@ contract('App', function ([_, owner, otherAdmin]) {
       });
 
       it('should register implementation on directory', async function () {
-        const implementation = await this.app.currentDirectory().getImplementation(contractName);
+        const implementation = await this.app.directory.getImplementation(contractName);
         implementation.should.eq(this.implementation_v1.address);
       });
     });
@@ -126,7 +130,7 @@ contract('App', function ([_, owner, otherAdmin]) {
 
       it('should unset implementation on directory', async function () {
         await this.app.unsetImplementation(contractName)
-        const implementation = await this.app.currentDirectory().getImplementation(contractName)
+        const implementation = await this.app.directory.getImplementation(contractName)
         implementation.should.be.zeroAddress
       })
     })
@@ -231,7 +235,7 @@ contract('App', function ([_, owner, otherAdmin]) {
 
     describe('setStdlib', function () {
       beforeEach('setting stdlib from name', async function () {
-        await this.app.setStdlib(stdlibAddress);
+        await this.app.setStdlib(this.stdlib.address);
       });
 
       shouldConnectToStdlib();
@@ -252,7 +256,7 @@ contract('App', function ([_, owner, otherAdmin]) {
 
   describe('with stdlib', function () {
     beforeEach('deploying', async function () {
-      this.app = await App.deployWithStdlib(initialVersion, stdlibAddress, txParams);
+      this.app = await App.deployWithStdlib(initialVersion, this.stdlib.address, txParams);
     });
 
     describe('deploy', function () {
@@ -260,9 +264,9 @@ contract('App', function ([_, owner, otherAdmin]) {
       shouldConnectToStdlib();
     });
 
-    describe('connect', function () {
+    describe('fetch', function () {
       beforeEach('connecting to existing instance', async function () {
-        this.app = await App.fetch(this.app.address(), txParams);
+        this.app = await App.fetch(this.app.address, txParams);
       });
 
       shouldInitialize();