+++ packages/lib/changelog.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Major Unreleased]
+
+### Added
+- New freezable implementation directory wrapper
+- New base implementation directory wrapper
+- Address getter function for custom contract wrappers
+
+### Changed
+- Refactor directory wrapper to represent the real implementation directories hierarchy
+- Refactor package wrapper allowing to work with any directory type
+- Refactor `App` and `Package` wrappers to work with new directory wrapper models
+- Index event arguments to improve querying
+
+### Removed
+- Remove `Release` contract and wrapper
+
 ## v1.4.1 - 2018-08-22
 
 ### Fixed
@@ -11,6 +27,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Retry transactions up to 3 times upon "nonce too low" errors ([#334](https://github.com/zeppelinos/zos-cli/issues/334))
 
 ## v1.4.0 - 2018-08-14
+
+### Added
+- Support to allow creating non-upgradeable instances of registered contracts
 
 ### Changed
 - Contract deployments and transactions are executed with an estimate of the gas needed, instead of using the network default ([#211](https://github.com/zeppelinos/zos-lib/pull/211))